src/debugging_cake_portal/requirements.txt
asgiref==3.5.2
attrs==19.3.0
Automat==0.8.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
distlib==0.3.5
distro==1.4.0
Django==4.1
django-crispy-forms==1.14.0
django-etc==1.3.1
django-filter==22.1
django-hitcount==1.3.5
djangorestframework==3.13.1
dnspython==1.16.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.7.1
future==0.18.2
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
incremental==16.10.1
iniconfig==1.1.1
jeepney==0.8.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
paramiko==2.6.0
pbr==5.10.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.2
pluggy==1.0.0
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.7.0
pulsemixer==1.5.0
py==1.11.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pytest==7.1.3
python-dateutil==2.7.3
python-xlib==0.23
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
sqlparse==0.2.4
ssh-import-id==5.10
testresources==2.0.1
tomli==2.0.1
Twisted==18.9.0
urllib3==1.25.8
virtualenv==20.16.3
wadllib==1.3.3
zope.interface==4.7.1
